+++ src/maicos/modules/temperatureplanar.py
@@ -50,6 +50,7 @@ class TemperaturePlanar(ProfilePlanarBase):
         sym: bool = False,
         grouping: str = "atoms",
         unwrap: bool = True,
+        pack: bool = True,
         bin_method: str = "com",
         output: str = "temperature.dat",
         concfreq: int = 0,
@@ -62,6 +63,7 @@ class TemperaturePlanar(ProfilePlanarBase):
         super().__init__(
             atomgroup=atomgroup,
             unwrap=unwrap,
+            pack=pack,
             jitter=jitter,
             concfreq=concfreq,
             dim=dim,